requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
fsspec==2023.9.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
zipp==3.17.0
